+++ api/src/controllers/exports.controller.ts
@@ -17,6 +17,7 @@ import { yamlFlatExporter } from '../formatters/yaml-flat';
 import { yamlNestedExporter } from '../formatters/yaml-nested';
 import AuthorizationService from '../services/authorization.service';
 import { gettextExporter } from '../formatters/gettext';
+import { stringsExporter } from '../formatters/strings';
 
 @Controller('api/v1/projects/:projectId/exports')
 export class ExportsController {
@@ -92,6 +93,8 @@ export class ExportsController {
         return await propertiesExporter(data);
       case 'po':
         return await gettextExporter(data);
+      case 'strings':
+        return await stringsExporter(data);
       default:
         throw new Error('Export format not implemented');
     }