+++ api/src/controllers/import.controller.ts
@@ -32,6 +32,7 @@ import { yamlFlatParser } from '../formatters/yaml-flat';
 import { yamlNestedParser } from '../formatters/yaml-nested';
 import AuthorizationService from '../services/authorization.service';
 import { gettextParser } from '../formatters/gettext';
+import { stringsParser } from '../formatters/strings';
 
 @Controller('api/v1/projects/:projectId/imports')
 export class ImportController {
@@ -178,6 +179,8 @@ export class ImportController {
           return await propertiesParser(contents);
         case 'po':
           return await gettextParser(contents);
+        case 'strings':
+          return await stringsParser(contents);
         default:
           throw new Error('Export format not implemented');
       }