+++ api/src/domain/http.ts
@@ -170,6 +170,7 @@ export enum ImportExportFormat {
   YamlNested = 'yamlnested',
   Properties = 'properties',
   Gettext = 'po',
+  Strings = 'strings',
 }
 
 export class ExportQuery {